// src/types/LocationName.ts
export enum LocationName {
	AGRABAH = "Agrabah",
	ATLANTICA = "Atlantica",
	BEASTS_CASTLE = "Beast's Castle",
	CAVERN_OF_REMEMBRANCE = "Cavern of Remembrance",
	CRITICAL = "Critical Extra",
	DISNEY_CASTLE = "Disney Castle",
	DONALD = "Donald",
	FORM = "Form",
	GOOFY = "Goofy",
	HALLOWEEN_TOWN = "Halloween Town",
	HOLLOW_BASTION = "Hollow Bastion",
	LAND_OF_DRAGONS = "Land of Dragons",
	LEVEL_UP = "Level Up",
	OLYMPUS = "Olympus Colisseum",
	OLYMPUS_CUPS = "Olympus Colisseum (Cups)",
	ABSENT_SILHOUETTE = "Absent Silhouette",
	DATA_ORGANIZATION_XIII = "Data Organization XIII",
	POOH = "100 Acre Wood",
	PORT_ROYAL = "Port Royal",
	PRIDE_LANDS = "Pride Lands",
	SIMULATED_TWILIGHT_TOWN = "Simulated Twilight Town",
	SPACE_PARANOIDS = "Space Paranoids",
	SYNTHESIS = "Synthesis",
	TIMELESS_RIVER = "Timeless River",
	TWILIGHT_TOWN = "Twilight Town",
	TWTNW = "The World That Never Was",
}